Устранить замечания в ReviewService. Устранить замечание по log.info("Друг успешно добавлен").

+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -75,7 +75,7 @@ public class UserController {
 
     @GetMapping("/{id}/feed")
     public List<Event> getFeed(@PathVariable int id) {
-        return eventService.getFeed(id);
+        return eventService.getFeed((long) id);
     }
 
     @GetMapping("/{id}")
